Fix to displayed correct sub on UI
In apm mode, when emergency call is initiated UI is getting updated
with the user selected sub. But if the other sub got activated,
call will be initiated through the active sub. Then UI was
not getting updated.
Updated the Connection object with the correct PhoneAccountHandle.
Same account should be updated to the UI.

Change-Id: If22561a715efc6a66059082043d52157b0b56b21
CRs-Fixed: 1061553